+++ weibo_spider/__main__.py
@@ -1,5 +1,5 @@
 from absl import app
 
-from .weibo_spider import main
+from .spider import main
 
 app.run(main)
